requirements.txt
streamlit==1.31.0
groq==0.4.0
python-dotenv==1.0.0
numpy~=1.26.4
protobuf<5,>=3.20
# opentelemetry-proto removed due to protobuf conflict